+++ CHANGELOG.md
@@ -51,6 +51,8 @@ the `configuration.konghq.com` API group.
 
 #### Added
 
+- Added validation for `Gateway` objects in the admission webhook
+  [#1946](https://github.com/Kong/kubernetes-ingress-controller/issues/1946)
 - [Feature Gates][k8s-fg] have been added to the controller manager in order to
   enable alpha/beta/experimental features and provide documentation about those
   features and their maturity over time. For more information see the